+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/pages/citizen/FAQs/FAQs.js
@@ -1,4 +1,4 @@
-import { BackButton, Header, Loader ,SearchIconSvg} from "@egovernments/digit-ui-components";
+import { BackButton, HeaderComponent, Loader ,SearchIconSvg} from "@egovernments/digit-ui-components";
 import React, { Fragment } from "react";
 import { useTranslation } from "react-i18next";
 import FaqComponent from "./FaqComponent";
@@ -24,7 +24,7 @@ const FAQsSection = ({ module }) => {
       <div className="faq-page">
         <BackButton style={{ marginLeft: "unset" }}></BackButton>
         <div style={{ marginBottom: "15px" }}>
-          <Header styles={{ marginLeft: "0px", paddingTop: "10px", fontSize: "32px" }}>{t("FAQ_S")}</Header>
+          <HeaderComponent styles={{ marginLeft: "0px", paddingTop: "10px", fontSize: "32px" }}>{t("FAQ_S")}</HeaderComponent>
         </div>
         <div className="faq-list">
           {moduleFaqs.map((faq, i) => (