+++ micro-ui/web/micro-ui-internals/packages/modules/workbench/src/pages/employee/MDMSAddV2.js
@@ -120,7 +120,20 @@ const MDMSAdd = ({ defaultFormData, updatesToUISchema, screenType = "add", onVie
       return localModuleName.replace(/[^a-zA-Z0-9]/g, "-").toUpperCase();
   };
 
-  const onSubmit = (data, additionalProperties) => {
+  const schemaCodeToValidate = `${moduleName}.${masterName}`;
+  const onSubmit = async (data, additionalProperties) => {
+    const validationConfig = Digit?.Customizations?.["commonUiConfig"]?.["AddMdmsConfig"]?.[schemaCodeToValidate];
+    if (!validationConfig?.validateForm){
+      console.warn(`No validation configuration found for schema: ${schemaCodeToValidate}`);
+      return true;
+    }
+    const validation = await validationConfig.validateForm(data, { tenantId: tenantId });
+    if (validation && !validation?.isValid) {
+      setShowToast(t(validation.message) || t('VALIDATION_ERROR_DEFAULT'));
+      setShowErrorToast(true);
+      toggleSpinner(false);
+      return;
+    }
     let locale = Digit.StoreData.getCurrentLanguage();
     toggleSpinner(true);