+++ micro-ui/web/micro-ui-internals/packages/libraries/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+
+## [1.8.9]  [1-Feb-2025]
+- FEATURE/HCMPRE-1425 : Added the workbench module patches and Updated localisation search screen, and core module #2181
+- Upgraded with new Components in core, workbench screens
+
 ## [1.8.8]  [21-Jan-2025]
 - Removed support for any new context path to have employee linked in the url.To use this,the new context path should be linked with employee.